Use common MQTT code in MKR1000.
